+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 Supports pytorch 1.8
 ### Added
+- `iou_thresholds` parameter to `COCOMetric`
 - `SimpleConfusionMatrix` Metric
+
 ### Changed
 - **Breaking:** Albumentations `aug_tfms` defaults.
   - rotate_limit changed from 45 to 15